services/glossary_service/pyproject.toml
[tool.poetry]
name = "wbfw109-glossary-service"
version = "0.1.0"
description = ""
authors = ["wbfw109 <wbfw109@gmail.com>"]
packages = [
    { include= "wbfw109/glossary_app", from = "src"}
]


[tool.poetry.dependencies]
python = "~3.11, <3.12"

[tool.poetry.group.essential.dependencies]
study-core = {path = "../../../study-core", develop = true}

[tool.poetry.group.dev.dependencies]
black = "*"
bandit = "*"
pylint = "*"
pyclean = "^2.2.0"
autoflake = "^2.0.0"
pydocstringformatter = "^0.7.2"
pyupgrade = "^3.3.1"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-cov = "*"
pytest-bdd = "*"
pytest-selenium = "^4.0.0"
pytest-factoryboy = "^2.5.1"


[tool.poetry.group.framework.dependencies]
pynecone-io = {version = "^0.1.13", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
